+++ rio_tiler/io/base.py
@@ -358,9 +358,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         """Validate asset name and construct url."""
         ...
 
-    def _get_reader(self, asset_info: AssetInfo) -> tuple[type[BaseReader], dict]:
+    def _get_reader(self, asset_info: AssetInfo) -> type[BaseReader]:
         """Get Asset Reader and options."""
-        return self.reader, {}
+        return self.reader
 
     def _update_statistics(
         self,
@@ -403,12 +403,13 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         def _reader(asset: str, **kwargs: Any) -> dict:
             asset_info = self._get_asset_info(asset)
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    return src.info(**asset_info["method_options"])
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
+                    return src.info(**method_options)
 
         return multi_values(assets, _reader, **kwargs)
 
@@ -437,12 +438,12 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         def _reader(asset: str, *args: Any, **kwargs: Any) -> dict:
             asset_info = self._get_asset_info(asset)
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    method_options = {**asset_info["method_options"], **kwargs}
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
                     return src.statistics(*args, **method_options)
 
         return multi_values(assets, _reader, **kwargs)
@@ -547,12 +548,12 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
             asset_info = self._get_asset_info(asset)
             asset_name = asset_info["name"]
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    method_options = {**asset_info["method_options"], **kwargs}
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
                     data = src.tile(*args, **method_options)
 
                     self._update_statistics(
@@ -627,12 +628,12 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
             asset_info = self._get_asset_info(asset)
             asset_name = asset_info["name"]
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    method_options = {**asset_info["method_options"], **kwargs}
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
                     data = src.part(*args, **method_options)
 
                     self._update_statistics(
@@ -705,12 +706,12 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         def _reader(asset: str, **kwargs: Any) -> ImageData:
             asset_info = self._get_asset_info(asset)
             asset_name = asset_info["name"]
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    method_options = {**asset_info["method_options"], **kwargs}
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
                     data = src.preview(**method_options)
 
                     self._update_statistics(
@@ -787,12 +788,12 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         def _reader(asset: str, *args: Any, **kwargs: Any) -> PointData:
             asset_info = self._get_asset_info(asset)
             asset_name = asset_info["name"]
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    method_options = {**asset_info["method_options"], **kwargs}
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
                     data = src.point(*args, **method_options)
 
                     metadata = data.metadata or {}
@@ -862,12 +863,12 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
             asset_info = self._get_asset_info(asset)
             asset_name = asset_info["name"]
-            reader, reader_options = self._get_reader(asset_info)
+            reader = self._get_reader(asset_info)
+            reader_options = {**self.reader_options, **asset_info["reader_options"]}
+            method_options = {**asset_info["method_options"], **kwargs}
 
-            options = {**self.reader_options, **reader_options}
             with self.ctx(**asset_info.get("env", {})):
-                with reader(asset_info["url"], tms=self.tms, **options) as src:
-                    method_options = {**asset_info["method_options"], **kwargs}
+                with reader(asset_info["url"], tms=self.tms, **reader_options) as src:
                     data = src.feature(*args, **method_options)
 
                     self._update_statistics(